Refactor: InteractionResponse Subclasses
- Create InteractionResponse subclasses for each Interaction subclass
- Removed unnecessary imports from interaction.Base file
- Edited each method and class description to match the Interaction/-Response subclass
- Altered TypeHints to use the Relevant InteractionResponse subclasses
